Allocator interface chages to take container's name to get/update/create/remove page. This name may be used by allocator to optimize allocation locality.

+++ include/memoria/allocators/inmem/factory.hpp
@@ -47,7 +47,7 @@ typedef memoria::InMemAllocator<
 template <typename CtrName>
 using SCtrTF = CtrTF<SmallProfile<>, CtrName>;
 
-typedef PageID<UInt> ID4;
+typedef PageID<BigInt> ID4;
 
 MEMORIA_EXTERN_TREE(BigInt, BigInt,     1);
 MEMORIA_EXTERN_TREE(BigInt, EmptyValue, 1);
